src/main/java/com/profitbricks/rest/domain/Location.java
/*
 * Copyright 2015.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package com.profitbricks.rest.domain;

import java.util.ArrayList;
import java.util.List;

/**
 * @author jasmin.gacic
 */
public class Location extends ProfitbricksBase {

    /**
     * @return the properties
     */
    public Location.Properties getProperties() {
        return properties;
    }

    /**
     * @param properties the properties to set
     */
    public void setProperties(Properties properties) {
        this.properties = properties;
    }

    private Location.Properties properties = new Properties();

    public class Properties {
        private List<String> features;
        private String name;

        public List<String> getFeatures() {
            if (features == null) {
                features = new ArrayList<String>();
            }
            return features;
        }

        public void setFeatures(List<String> features) {
            this.features = features;
        }

        public String getName() {
            return name;
        }

        public void setName(String name) {
            this.name = name;
        }
    }

}
